requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.1
transformers==4.45.1
torch==2.5.0
pydantic==2.9.2
python-dotenv==1.0.1
